# swagger/openapi.yml
openapi: 3.1.0
info:
  title: Semantic Cuisine
  description: |-
    This is the API specification for the Semantic Browse for Cuisines application.
  version: 1.0.3
...
servers:
  - url: http://localhost:5173/api/v1
    description: Development compose
  - url: http://localhost:80/api/v1
    description: Production-like compose
  - url: https://semantic-cuisine-wh4f2.ondigitalocean.app/api/v1
    description: Production
  - url: https://semantic-cuisine-staging-t3m52.ondigitalocean.app/api/v1
    description: Staging
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags:
        - user
      summary: Get own profile
      security:
        - auth_jwt: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseObject"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserProfile"
  /users/{userId}/following:
    get:
      operationId: getUserFollowing
      tags:
        - user
      summary: Get users being followed
      security:
        - auth_jwt: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseObject"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserArray"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
  /users/{userId}/followers:
    get:
      operationId: getUserFollowers
      tags:
        - user
      summary: Get user's followers
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseObject"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserArray"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
  /users/{userId}/follow:
    post:
      operationId: followUser
      tags:
        - user
      summary: Follow a user
      security:
        - auth_jwt: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK - returning new user profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseObject"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
    delete:
      operationId: unfollowUser
      tags:
        - user
      summary: Unfollow a user
      security:
        - auth_jwt: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK - returning new user profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseObject"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /search/users:
    get:
      operationId: searchUsers
      tags:
        - user
      summary: Search for users
      parameters:
        - in: query
          name: q
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseObject"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserArray"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
  /search/dishes:
    get:
      operationId: searchDishes
      tags:
        - dish
      summary: Search for dishes
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: cuisine
          schema:
            type: string
            description: Comma separated list of cuisines
        - in: query
          name: foodType
          schema:
            type: string
            description: Comma separated list of food types
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DishArray"
                  - $ref: "#/components/schemas/SuccessResponseObject"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
...
  /cuisines/{cuisineId}:
    get:
      operationId: getCuisineById
      tags:
        - cuisines
      summary: Get cuisine details
      parameters:
        - in: path
          name: cuisineId
          required: true
          schema:
            type: integer
        - in: query
          name: includeDishes
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseObject"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CuisineDetails"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
...
  /recipes:
    get:
      operationId: getRecipesForEntity
      tags:
        - recipe
      summary: Get recipes for dish, cuisine
      parameters:
        - in: query
          name: sort
          schema:
            type: string
            enum: [recent, topRated]
        - in: query
          name: dishId
          schema:
            type: integer
        - in: query
          name: cuisineId
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseObject"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RecipeArray"
        "400":
          $ref: "#/components/responses/BadRequestResponse"

    post:
      operationId: createRecipe
      tags:
        - recipe
      summary: Create a new recipe
      security:
        - auth_jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewRecipe"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseObject"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RecipeDetails"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
  /recipes/{recipeId}:
    get:
      operationId: getRecipeById
      tags:
        - recipe
      summary: Get recipe details
      parameters:
        - in: path
          name: recipeId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseObject"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RecipeDetails"
        "404":
          $ref: "#/components/responses/NotFoundResponse"

    delete:
      operationId: deleteRecipeById
      tags:
        - recipe
      summary: Delete a recipe
      security:
        - auth_jwt: []
      parameters:
        - in: path
          name: recipeId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: OK
          $ref: "#/components/responses/OkResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /recipes/{recipeId}/rating:
    post:
      operationId: rateRecipe
      tags:
        - recipe
      summary: Rate a recipe
      security:
        - auth_jwt: []
      parameters:
        - in: path
          name: recipeId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
      responses:
        "200":
          $ref: "#/components/responses/OkResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        bio:
          type: string
        followersCount:
          type: integer
        gender:
          type: string
          enum: [male, female, unknown]
        profilePicture:
          type: string
          format: uri
        diets:
          type: array
          items:
            type: string
        recipeCount:
          type: integer
        bookmarks:
          type: array
          description: Only available when querying the current user's profile.
          items:
            $ref: "#/components/schemas/RecipeSummary"
        recipes:
          type: array
          items:
            $ref: "#/components/schemas/RecipeSummary"
      examples:
        - id: 1
          username: "takoyaki_lover"
          name: "Takoyaki Lover"
          bio: "I love takoyaki!"
          followersCount: 100
          gender: unknown
          profilePicture: "https://images.unsplash.com/photo-1633790450512-98e68a55ef15?ixlib=rb-4.0.3&q=85&fm=jpg&crop=entropy&cs=srgb&dl=brunno-tozzo-GAIC2WHxm5A-unsplash.jpg&w=640"
          diets: ["keto"]
          recipeCount: 10
          bookmarks:
            - id: 1
              name: "My Takoyaki Recipe"
              description: "A delicious takoyaki recipe that I learned from my grandmother."
              cookTime: 30
              images:
                - "http://commons.wikimedia.org/wiki/Special:FilePath/Takoyaki%20by%20yomi955.jpg"
              rating: 4.5
              dish:
                id: "http://www.wikidata.org/entity/Q905527"
                name: "takoyaki"
          recipes:
            - id: 1
              name: "My Takoyaki Recipe"
              description: "A delicious takoyaki recipe that I learned from my grandmother."
              cookTime: 30
              images:
                - "http://commons.wikimedia.org/wiki/Special:FilePath/Takoyaki%20by%20yomi955.jpg"
              rating: 4.5
              dish:
                id: "http://www.wikidata.org/entity/Q905527"
                name: "takoyaki"
    UserSummary:
      type: object
      required:
        - id
        - username
        - name
        - followersCount
        - profilePicture
        - recipeCount
        - avgRating
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        followersCount:
          type: integer
        profilePicture:
          type: string
          format: uri
        recipeCount:
          type: integer
        avgRating:
          type: number
          format: float
      examples:
        - id: 1
          username: "takoyaki_lover"
          name: "Takoyaki Lover"
          followersCount: 100
          profilePicture: "http://commons.wikimedia.org/wiki/Special:FilePath/Takoyaki%20by%20yomi955.jpg"
          recipeCount: 10
          avgRating: 4.5
    DishDetails:
      type: object
      examples:
        - id: "http://www.wikidata.org/entity/Q905527"
          name: "takoyaki"
          description: "ball-shaped Japanese snack with octopus"
          image: "http://commons.wikimedia.org/wiki/Special:FilePath/Takoyaki%20by%20yomi955.jpg"
          countries: "Japan"
          ingredients: "wheat flour, green laver, octopus as food, beni shōga, katsuobushi, Welsh onion"
          foodTypes: "konamono, octopus dish, yakimono"
          cuisines: "Japanese cuisine"
      required:
        - id
        - name
        - image
        - description
        - countries
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        image:
          type: string
          format: uri
        countries:
          type: string
        ingredients:
          type: string
        foodTypes:
          type: string
        cuisine:
          $ref: "#/components/schemas/CuisineSummary"
        recipes:
          type: array
          description: Only returned when directly querying a dish.
          items:
            $ref: "#/components/schemas/RecipeSummary"
    CuisineDetails:
      type: object
      required:
        - id
        - name
        - description
        - image
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        image:
          type: string
          format: uri
        isSelfFollowing:
          type: boolean
        dishes:
          type: array
          items:
            $ref: "#/components/schemas/DishSummary"
    RecipeSummary:
      type: object
      required:
        - id
        - name
        - description
        - images
        - rating
        - cookTime
        - avgRating
        - ratingsCount
        - author
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        cookTime:
          type: integer
          description: Cook time in minutes
        images:
          type: array
          items:
            type: string
            format: uri
        avgRating:
          type: number
          format: float
        ratingsCount:
          type: integer
        selfRating:
          type: integer
        dish:
          $ref: "#/components/schemas/DishSummary"
        author:
          $ref: "#/components/schemas/UserSummary"
      examples:
        - id: 1
          name: "My Takoyaki Recipe"
          description: "A delicious takoyaki recipe that I learned from my grandmother."
          cookTime: 30
          images:
            - "http://commons.wikimedia.org/wiki/Special:FilePath/Takoyaki%20by%20yomi955.jpg"
          avgRating: 4.5
          ratingsCount: 42
          selfRating: 5
          dish:
            id: "http://www.wikidata.org/entity/Q905527"
            name: "takoyaki"
            image: "http://commons.wikimedia.org/wiki/Special:FilePath/Takoyaki%20by%20yomi955.jpg"
          author:
            id: 1
            username: "takoyaki_lover"
            name: "Takoyaki Lover"
            followersCount: 100
            profilePicture: "http://commons.wikimedia.org/wiki/Special:FilePath/Takoyaki%20by%20yomi955.jpg"
            recipeCount: 10
            avgRating: 4.5
    RecipeDetails:
      type: object
      required:
        - id
        - name
        - description
        - instructions
        - ingredients
        - images
        - cookTime
        - servingSize
        - allergens
        - ratingsCount
        - author
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        instructions:
          type: array
          items:
            type: string
        ingredients:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
            format: uri
        cookTime:
          type: string
        servingSize:
          type: integer
        allergens:
          type: array
          items:
            type: string
        cuisine:
          $ref: "#/components/schemas/CuisineSummary"
        dish:
          $ref: "#/components/schemas/DishSummary"
        avgRating:
          type: number
          format: float
        ratingsCount:
          type: integer
        selfRating:
          type: integer
          description: The current user's rating for this recipe, if any.
        author:
          $ref: "#/components/schemas/UserSummary"
      examples:
        - id: 1
          name: "My Takoyaki Recipe"
          description: "A delicious takoyaki recipe that I learned from my grandmother."
          instructions:
            - "Mix the batter."
            - "Add the octopus."
            - "Cook the takoyaki balls."
          ingredients:
            - "wheat flour"
            - "green laver"
            - "octopus as food"
            - "beni shōga"
            - "katsuobushi"
            - "Welsh onion"
          images:
            - "http://commons.wikimedia.org/wiki/Special:FilePath/Takoyaki%20by%20yomi955.jpg"
          cookTime: "30 minutes"
          servingSize: 4
          allergens:
            - "seafood"
          cuisine:
            id: 1
            name: "Japanese"
          dish:
            id: "http://www.wikidata.org/entity/Q905527"
            name: "takoyaki"
            image: "http://commons.wikimedia.org/wiki/Special:FilePath/Takoyaki%20by%20yomi955.jpg"
          avgRating: 4.5
          ratingsCount: 10
    NewRecipe:
      type: object
      required:
        - name
        - instructions
        - ingredients
        - prepTime
        - cookTime
        - servingSize
      properties:
        name:
          type: string
        description:
          type: string
        instructions:
          type: array
          items:
            type: string
        ingredients:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
            format: binary
        prepTime:
          type: string
        cookTime:
          type: string
        servingSize:
          type: integer
        allergens:
          type: array
          items:
            type: string
        dishId:
          type: integer
      examples:
        - name: "My Takoyaki Recipe"
          description: "A delicious takoyaki recipe that I learned from my grandmother."
          instructions:
            - "Mix the batter."
            - "Add the octopus."
            - "Cook the takoyaki balls."
          ingredients:
            - "wheat flour"
            - "green laver"
            - "octopus as food"
            - "beni shōga"
            - "katsuobushi"
            - "Welsh onion"
          images:
            - "http://commons.wikimedia.org/wiki/Special:FilePath/Takoyaki%20by%20yomi955.jpg"
          prepTime: "10 minutes"
          cookTime: "30 minutes"
          servingSize: 4
          allergens:
            - "seafood"
          dishId: 1
    CuisineSummary:
      type: object
      required: ["id", "name"]
      properties:
        id:
          type: integer
        name:
          type: string
      examples:
        - id: 1
          name: "Japanese"
    DishSummary:
      type: object
      required: ["id", "name"]
      properties:
        id:
          type: integer
        name:
          type: string
      examples:
        - id: 1
          name: "takoyaki"
...
    DishArray:
      description: An array of dishes
      type: array
      items:
        $ref: "#/components/schemas/DishDetails"
    RecipeArray:
      description: An array of recipes
      type: array
      items:
        $ref: "#/components/schemas/RecipeDetails"
    UserArray:
      description: An array of users
      type: array
      items:
        $ref: "#/components/schemas/UserSummary"
...
    SuccessResponseObject:
      description: OK
      type: object
      required: ["status", "data"]
      properties:
        status:
          type: integer
          description:
            Internal status code of the response. An HTTP 200 response with an internal
            500 status code is an error response. Prioritize the inner status over the HTTP
            status.
          enum:
            - 200
            - 201
          examples: [200, 201]
        data:
          oneOf:
            - type: object
            - type: array
    ErrorResponseObject:
      description: Response with errors
      type: object
      required: ["status", "errors"]
      properties:
        status:
          enum:
            - 400
            - 401
            - 403
            - 404
            - 409
            - 500
          description:
            Internal status code of the response. An HTTP 200 response with an internal
            500 status code is an error response. Prioritize the inner status over the HTTP
            status.
          examples: [400, 401, 403, 404, 409, 500]
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiError"
  responses:
    OkResponse:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessResponseObject"
...
    BadRequestResponse:
      description: Bad Request
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorResponseObject"
              - type: object
                properties:
                  status:
                    const: 400
...
    ForbiddenResponse:
      description: Forbidden
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorResponseObject"
              - type: object
                properties:
                  status:
                    const: 403
    NotFoundResponse:
      description: Not Found
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorResponseObject"
              - type: object
                properties:
                  status:
                    const: 404
...
  securitySchemes:
    auth_jwt:
      type: http
      description: A bearer token is used to authenticate a user.
      scheme: bearer
      bearerFormat: JWT